Refactor archive package, fix failing tests
